GH-1016 Improvements
