+++ front-end/src/routes/MainRoutes.tsx
@@ -16,13 +16,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 import { lazy } from "react";
-import { useSelector } from "react-redux";
 import { Navigate, Route, Routes } from "react-router";
 
 import Loadable from "components";
+import { useAppSelector } from "hooks";
 import { MainLayout } from "layout";
-
-import type { RootState } from "../store";
 
 const GarbageCollector = Loadable(lazy(() => import("pages/GarbageCollector")));
 const ScheduledTasks = Loadable(lazy(() => import("pages/ScheduledTasks")));
@@ -41,7 +39,7 @@ const Beans = Loadable(lazy(() => import("pages/Beans")));
 const MCP = Loadable(lazy(() => import("pages/MCP")));
 
 export const MainRoutes = () => {
-    const settings = useSelector((state: RootState) => state.settings);
+    const settings = useAppSelector((state) => state.settings);
 
     return (
         <>
