+++ front-end/src/services/index.ts
@@ -23,6 +23,7 @@ export * from "./threadDump";
 export * from "./conditions";
 export * from "./wallboard";
 export * from "./dashboard";
+export * from "./settings";
 export * from "./loggers";
 export * from "./details";
 export * from "./metrics";
@@ -30,4 +31,3 @@ export * from "./caches";
 export * from "./beans";
 export * from "./auth";
 export * from "./gc";
-export * from "./settings";
