+++ front-end/src/store/slices/index.ts
@@ -16,3 +16,4 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 export * from "./authoritiesSlice";
+export * from "./settingsSlice";
